chore: re-organize funcs
