+++ src/main/consts.ts
@@ -1,0 +1,9 @@
+import { app } from "electron";
+import { join } from "node:path";
+
+export const WORKING_DIR = join(__dirname, "../../");
+export const DIST_ELECTRON = join(WORKING_DIR, "out");
+export const PUBLIC_DIR = join(
+  WORKING_DIR,
+  app.isPackaged ? "../public" : "public",
+);
