+++ src/main/index.ts
@@ -1,16 +1,7 @@
-import {
-  app,
-  BrowserWindow,
-  shell,
-  ipcMain,
-  nativeImage,
-  dialog,
-} from "electron";
+import { app, BrowserWindow, ipcMain, dialog } from "electron";
 import contextMenu from "electron-context-menu";
 import { release } from "node:os";
-import { join } from "node:path";
-import { checkForUpdates, update } from "./update";
-import { ChildProcess } from "node:child_process";
+import { checkForUpdates } from "./update";
 import log from "electron-log/main";
 import { API } from "../types/api";
 import {
@@ -35,13 +26,12 @@ import {
   cacheCustomBlocksDir,
   getCustomBlocksDir,
   ifconfig,
-  isPortFree,
-  killProcess,
   netstat,
   pickDirectory,
   ping,
   openFilePicker,
   writeFileSync,
+  cleanup,
 } from "./utils";
 import {
   browsePythonInterpreter,
@@ -53,6 +43,7 @@ import {
   poetryShowTopLevel,
   poetryUninstallDepGroup,
 } from "./python/poetry";
+import { createWindow } from "./window";
 
 log.initialize({ preload: true });
 log.info("Welcome to Flojoy Studio!");
@@ -73,9 +64,6 @@ const envPath = process.env.PATH ?? "";
 if (!envPath.split(":").includes("usr/local/bin")) {
   process.env.PATH = [...envPath.split(":"), "usr/local/bin"].join(":");
 }
-const WORKING_DIR = join(__dirname, "../../");
-const DIST_ELECTRON = join(WORKING_DIR, "out");
-const PUBLIC_DIR = join(WORKING_DIR, app.isPackaged ? "../public" : "public");
 
 // Disable GPU Acceleration for Windows 7
 if (release().startsWith("6.1")) app.disableHardwareAcceleration();
@@ -87,17 +75,6 @@ if (!app.requestSingleInstanceLock()) {
   app.quit();
   process.exit(0);
 }
-
-const getIcon = () => {
-  switch (process.platform) {
-    case "win32":
-      return join(PUBLIC_DIR, "favicon.ico");
-    case "linux":
-      return join(PUBLIC_DIR, "favicon.png");
-    default:
-      return join(PUBLIC_DIR, "favicon.png");
-  }
-};
 
 const handleSetUnsavedChanges = (_, value: boolean) => {
   global.hasUnsavedChanges = value;
@@ -135,87 +112,7 @@ contextMenu({
   },
 });
 
-// Here, you can also use other preload
-const preload = join(__dirname, `../preload/index.js`);
-
-const devServerUrl = process.env["ELECTRON_RENDERER_URL"];
-const indexHtml = join(DIST_ELECTRON, "renderer", "index.html");
-
 app.setName("Flojoy Studio");
-
-async function createWindow() {
-  const mainWindow = new BrowserWindow({
-    title: "Flojoy Studio",
-    icon: getIcon(),
-    autoHideMenuBar: true,
-    titleBarStyle: process.platform === "darwin" ? "hidden" : "default",
-    trafficLightPosition: {
-      x: 15,
-      y: 17, // macOS traffic lights seem to be 14px in diameter. If you want them vertically centered, set this to `titlebar_height / 2 - 7`.
-    },
-    webPreferences: {
-      preload,
-      sandbox: false,
-    },
-    show: false,
-  });
-  global.mainWindow = mainWindow;
-  global.hasUnsavedChanges = true;
-
-  mainWindow.on("ready-to-show", () => {
-    mainWindow.show();
-    mainWindow.maximize();
-  });
-  // setting icon for mac
-  if (process.platform === "darwin") {
-    app.dock.setIcon(nativeImage.createFromPath(getIcon()));
-  }
-  if (!(await isPortFree(5392))) {
-    const choice = dialog.showMessageBoxSync(global.mainWindow, {
-      type: "question",
-      buttons: ["Exit", "Kill Process"],
-      title: "Existing Server Detected",
-      message:
-        "Seems like there is already a Flojoy server running! Do you want to kill it?",
-      icon: getIcon(),
-    });
-    if (choice == 0) {
-      mainWindow.destroy();
-      app.quit();
-      process.exit(0);
-    } else {
-      await killProcess(5392).catch((err) => log.error(err));
-    }
-  }
-  if (!app.isPackaged && devServerUrl) {
-    await mainWindow.loadURL(devServerUrl);
-  } else {
-    await mainWindow.loadFile(indexHtml);
-  }
-
-  mainWindow.on("close", (e) => {
-    if (!global.hasUnsavedChanges) {
-      return;
-    }
-    const choice = dialog.showMessageBoxSync(global.mainWindow, {
-      type: "question",
-      buttons: ["Yes", "No, go back"],
-      title: "Quit?",
-      message:
-        "You have unsaved changes. Are you sure you want to quit Flojoy Studio?",
-    });
-    if (choice > 0) e.preventDefault();
-  });
-
-  // Make all links open with the browser, not with the application
-  mainWindow.webContents.setWindowOpenHandler(({ url }) => {
-    if (url.startsWith("https:")) shell.openExternal(url);
-    return { action: "deny" };
-  });
-
-  // Apply electron-updater
-  update(cleanup);
-}
 
 app.whenReady().then(async () => {
   createWindow().catch((err) => console.log(err));
@@ -296,35 +193,3 @@ app.on("activate", () => {
     createWindow();
   }
 });
-
-// New window example arg: new windows url
-ipcMain.handle("open-win", (_, arg) => {
-  const childWindow = new BrowserWindow({
-    webPreferences: {
-      preload,
-    },
-  });
-
-  if (process.env.VITE_DEV_SERVER_URL) {
-    childWindow.loadURL(`${devServerUrl}#${arg}`);
-  } else {
-    childWindow.loadFile(indexHtml, { hash: arg });
-  }
-});
-
-const cleanup = async () => {
-  const captainProcess = global.captainProcess as ChildProcess;
-  log.info(
-    "Cleanup function invoked, is captain running? ",
-    !(captainProcess?.killed ?? true),
-  );
-  if (captainProcess && captainProcess.exitCode === null) {
-    const success = killCaptain();
-    if (success) {
-      global.captainProcess = null;
-      log.info("Successfully terminated captain :)");
-    } else {
-      log.error("Something went wrong when terminating captain!");
-    }
-  }
-};
