+++ src/main/window.ts
@@ -1,0 +1,98 @@
+import { app, BrowserWindow, shell, nativeImage, dialog } from "electron";
+import { update } from "./update";
+import log from "electron-log/main";
+import { cleanup, isPortFree, killProcess } from "./utils";
+
+import { join } from "node:path";
+import { DIST_ELECTRON, PUBLIC_DIR } from "./consts";
+
+const getIcon = () => {
+  switch (process.platform) {
+    case "win32":
+      return join(PUBLIC_DIR, "favicon.ico");
+    case "linux":
+      return join(PUBLIC_DIR, "favicon.png");
+    default:
+      return join(PUBLIC_DIR, "favicon.png");
+  }
+};
+
+// Here, you can also use other preload
+const preload = join(__dirname, `../preload/index.js`);
+
+const devServerUrl = process.env["ELECTRON_RENDERER_URL"];
+const indexHtml = join(DIST_ELECTRON, "renderer", "index.html");
+
+export async function createWindow() {
+  const mainWindow = new BrowserWindow({
+    title: "Flojoy Studio",
+    icon: getIcon(),
+    autoHideMenuBar: true,
+    titleBarStyle: process.platform === "darwin" ? "hidden" : "default",
+    trafficLightPosition: {
+      x: 15,
+      y: 17, // macOS traffic lights seem to be 14px in diameter. If you want them vertically centered, set this to `titlebar_height / 2 - 7`.
+    },
+    webPreferences: {
+      preload,
+      sandbox: false,
+    },
+    show: false,
+  });
+  global.mainWindow = mainWindow;
+  global.hasUnsavedChanges = true;
+
+  mainWindow.on("ready-to-show", () => {
+    mainWindow.show();
+    mainWindow.maximize();
+  });
+  // setting icon for mac
+  if (process.platform === "darwin") {
+    app.dock.setIcon(nativeImage.createFromPath(getIcon()));
+  }
+  if (!(await isPortFree(5392))) {
+    const choice = dialog.showMessageBoxSync(global.mainWindow, {
+      type: "question",
+      buttons: ["Exit", "Kill Process"],
+      title: "Existing Server Detected",
+      message:
+        "Seems like there is already a Flojoy server running! Do you want to kill it?",
+      icon: getIcon(),
+    });
+    if (choice == 0) {
+      mainWindow.destroy();
+      app.quit();
+      process.exit(0);
+    } else {
+      await killProcess(5392).catch((err) => log.error(err));
+    }
+  }
+  if (!app.isPackaged && devServerUrl) {
+    await mainWindow.loadURL(devServerUrl);
+  } else {
+    await mainWindow.loadFile(indexHtml);
+  }
+
+  mainWindow.on("close", (e) => {
+    if (!global.hasUnsavedChanges) {
+      return;
+    }
+    const choice = dialog.showMessageBoxSync(global.mainWindow, {
+      type: "question",
+      buttons: ["Yes", "No, go back"],
+      title: "Quit?",
+      message:
+        "You have unsaved changes. Are you sure you want to quit Flojoy Studio?",
+    });
+    if (choice > 0) e.preventDefault();
+  });
+
+  // Make all links open with the browser, not with the application
+  mainWindow.webContents.setWindowOpenHandler(({ url }) => {
+    if (url.startsWith("https:")) shell.openExternal(url);
+    return { action: "deny" };
+  });
+
+  // Apply electron-updater
+  update(cleanup);
+}
